// src/app/models/userInqury.ts
export class UserInqury {
        id?: number;
        employeeId?: number;
        employeeNO?: string;
        employeeNameSl?: string;
        employeeNameFl?: string;
        isActive?: boolean;
        isCivilId?: boolean;
        isAuthorizeLogs?:boolean;
        expireDate?: Date;       
        isEndOfContract?: boolean;      
        username?: string;
        password?: string;
        civilId?: string;
        userNameTypeId?: number;
        userNameTypeFl?: string;
        userNameTypeSl?: string;
}